feat: add cli to python package
