+++ README.md
@@ -14,6 +14,14 @@ The published package includes plugin.wasm for browser/CDN usage.
 - crates/beancount-formatter-cli: CLI wrapper around the formatter library.
 - crates/beancount-formatter-py: Python bindings built with PyO3/maturin.
 
+## CLI usage
+- Format and print to stdout: `beancount-format path/to/file.beancount`
+- Write changes in place: `beancount-format --in-place path/to/dir`
+- Check without modifying (exit 1 if reformat needed): `beancount-format --check path/to/file.beancount`
+- Override config: `--line-width 80 --indent-width 4 --new-line lf --prefix-width 60 --num-width 20`
+
+The CLI auto-detects `pyproject.toml` from the working directory or provided paths and merges settings with any CLI overrides.
+
 ## Development
 - Build CLI: `cargo build -p beancount-formatter-cli --release`
 - Run tests: `cargo test --workspace`
