+++ crates/beancount-formatter-py/src/lib.rs
@@ -2,6 +2,7 @@
 
 use beancount_formatter::configuration::{Configuration, NewLineKind};
 use beancount_formatter::format;
+use beancount_formatter_cli;
 use pyo3::exceptions::{PyRuntimeError, PyValueError};
 use pyo3::prelude::*;
 
@@ -39,8 +40,17 @@ fn format_text_py(
   Ok(formatted)
 }
 
+#[pyfunction(name = "main")]
+fn main_py(args: Vec<String>) -> PyResult<bool> {
+  let outcome = beancount_formatter_cli::main_with_args(args)
+    .map_err(|err| PyRuntimeError::new_err(err.to_string()))?;
+
+  Ok(outcome.any_changed)
+}
+
 #[pymodule]
 fn beancount_format(_py: Python<'_>, m: &Bound<'_, PyModule>) -> PyResult<()> {
   m.add_function(wrap_pyfunction!(format_text_py, m)?)?;
+  m.add_function(wrap_pyfunction!(main_py, m)?)?;
   Ok(())
 }
